Deleted the stuff we don’t need tuned up the things that bother me.
It feels good, it works good, it runs clean and it has everything we need for now. After thinking about this deep and hard, we’re keeping the ranch in tact, and just stripping off the stuff we don’t want anymore, or that does not fit with our mission. copyrights remain in place, license remains in place, but this ain’t transport no more.

Signed-off-by: Dave Shanley <dave@quobix.com>